feat: handle retrieval queries for unindexed identity payload CIDs
There are valid cases where a CAR may have an identity CID as its root that is
not represented as a 'block' within the CAR body and therefore isn't indexed
by the dagstore. In this case, we inspect the identity CID content and treat
the query as a query for the intersection of all of the links within the block.

Ref: https://github.com/filecoin-project/boost/pull/715